Create dotnet module.
Split the dotnet functionality off from the pe module and put it in a
separate module called dotnet.

The pieces which are specific only to the pe module now live in
yara/pe.h. These include IMPORTED_DLL and IMPORTED_FUNCTION.

The pieces which are specific only to the dotnet module now live in
yara/dotnet.h. These include CLI_HEADER, NET_METADATA, etc.

The common pieces now live in pe_utils.c and the declarations and
necessary macros live in yara/pe_utils.h

The dotnet module uses the PE structure defined in pe.h, even though all
of it isn't necessary.

Add include guard around yara/pe.h.
